Search Bar Updated
